Refactor pcapng: move DecryptionSecrets to its own file
