+++ src/pcapng.rs
@@ -48,6 +48,7 @@ use rusticata_macros::{align32, newtype_enum};
 use std::borrow::Cow;
 use std::convert::TryFrom;
 
+mod decryption_secrets;
 mod enhanced_packet;
 mod interface_description;
 mod interface_statistics;
@@ -56,6 +57,7 @@ mod section_header;
 mod simple_packet;
 mod systemd_journal_export;
 
+pub use decryption_secrets::*;
 pub use enhanced_packet::*;
 pub use interface_description::*;
 pub use interface_statistics::*;
@@ -287,69 +289,6 @@ pub fn build_ts_f64(ts_high: u32, ts_low: u32, ts_offset: u64, resolution: u64) 
     let ts_fractional = (ts % resolution) as u32;
     // XXX should we round to closest unit?
     ts_sec as f64 + ((ts_fractional as f64) / (resolution as f64))
-}
-
-#[derive(Clone, Copy, Eq, PartialEq)]
-pub struct SecretsType(pub u32);
-
-newtype_enum! {
-    impl debug SecretsType {
-        TlsKeyLog = 0x544c_534b, // TLSK
-        WireguardKeyLog = 0x5747_4b4c,
-    }
-}
-
-#[derive(Debug)]
-pub struct DecryptionSecretsBlock<'a> {
-    pub block_type: u32,
-    pub block_len1: u32,
-    pub secrets_type: SecretsType,
-    pub secrets_len: u32,
-    pub data: &'a [u8],
-    pub options: Vec<PcapNGOption<'a>>,
-    pub block_len2: u32,
-}
-
-impl<'a, En: PcapEndianness> PcapNGBlockParser<'a, En, DecryptionSecretsBlock<'a>>
-    for DecryptionSecretsBlock<'a>
-{
-    const HDR_SZ: usize = 20;
-    const MAGIC: u32 = DSB_MAGIC;
-
-    fn inner_parse<E: ParseError<&'a [u8]>>(
-        block_type: u32,
-        block_len1: u32,
-        i: &'a [u8],
-        block_len2: u32,
-    ) -> IResult<&'a [u8], DecryptionSecretsBlock<'a>, E> {
-        // caller function already tested header type(magic) and length
-        // read end of header
-        let (i, secrets_type) = En::parse_u32(i)?;
-        let (i, secrets_len) = En::parse_u32(i)?;
-        // read packet data
-        // align32 can overflow
-        if secrets_len >= u32::MAX - 4 {
-            return Err(Err::Error(E::from_error_kind(i, ErrorKind::Verify)));
-        }
-        let padded_length = align32!(secrets_len);
-        let (i, data) = take(padded_length)(i)?;
-        // read options
-        let current_offset = (20 + padded_length) as usize;
-        let (i, options) = opt_parse_options::<En, E>(i, block_len1 as usize, current_offset)?;
-        if block_len2 != block_len1 {
-            return Err(Err::Error(E::from_error_kind(i, ErrorKind::Verify)));
-        }
-        let block = DecryptionSecretsBlock {
-            block_type,
-            block_len1,
-            secrets_type: SecretsType(secrets_type),
-            secrets_len,
-            data,
-            options,
-            block_len2,
-        };
-        Ok((i, block))
-    }
 }
 
 #[derive(Debug)]
@@ -671,20 +610,6 @@ fn if_extract_tsoffset_and_tsresol(options: &[PcapNGOption]) -> (u8, i64) {
         }
     }
     (if_tsresol, if_tsoffset)
-}
-
-#[inline]
-pub fn parse_decryptionsecretsblock_le(
-    i: &[u8],
-) -> IResult<&[u8], DecryptionSecretsBlock, PcapError<&[u8]>> {
-    ng_block_parser::<DecryptionSecretsBlock, PcapLE, _, _>()(i)
-}
-
-#[inline]
-pub fn parse_decryptionsecretsblock_be(
-    i: &[u8],
-) -> IResult<&[u8], DecryptionSecretsBlock, PcapError<&[u8]>> {
-    ng_block_parser::<DecryptionSecretsBlock, PcapBE, _, _>()(i)
 }
 
 #[inline]
